extract reduce and parse modules from term
