+++ src/lambda/index.ts
@@ -1,0 +1,3 @@
+export {Identifier, Term, variable, abstraction, application} from './term';
+export {parse} from './parse';
+export {reduce} from './reduce';
